rename of truncation folder -> conversion + moved util functions out of script
