+++ src/conversion/cipher_truncator.py
