+++ src/conversion/truncation_config.py
